Refactored the whole codebase so it is split better between files.
